+++ packages/frontend/src/scripts/sims/common/potency_ratio.ts
@@ -1,5 +1,5 @@
-import {noSimSettings, SimResult, SimSettings, SimSpec, Simulation} from "../simulation";
-import {CharacterGearSet} from "../gear";
+import {noSimSettings, SimResult, SimSettings, SimSpec, Simulation} from "../../simulation";
+import {CharacterGearSet} from "../../gear";
 import {applyDhCrit, baseDamage} from "@xivgear/xivmath/xivmath";
 
 export const potRatioSimSpec: SimSpec<PotencyRatioSim, SimSettings> = {
