+++ packages/frontend/src/scripts/sims/default_sims.ts
@@ -1,18 +1,18 @@
-import {whmSheetSpec} from "./whm_sheet_sim";
-import {sgeSheetSpec} from "./sge_sheet_sim";
-import {sgeNewSheetSpec} from "./sge_sheet_sim_mk2";
-import {astNewSheetSpec} from "./ast_sheet_sim";
-import {schNewSheetSpec} from "./sch_sheet_sim";
-import {whmNewSheetSpec} from "./whm_new_sheet_sim";
-import {rprSheetSpec} from "./rpr_sheet_sim";
+import {whmSheetSpec} from "./healer/whm_sheet_sim";
+import {sgeSheetSpec} from "./healer/sge_sheet_sim";
+import {sgeNewSheetSpec} from "./healer/sge_sheet_sim_mk2";
+import {astNewSheetSpec} from "./healer/ast_sheet_sim";
+import {schNewSheetSpec} from "./healer/sch_sheet_sim";
+import {whmNewSheetSpec} from "./healer/whm_new_sheet_sim";
+import {rprSheetSpec} from "./melee/rpr_sheet_sim";
 import {registerSim} from "../simulation";
-import {potRatioSimSpec} from "./potency_ratio";
-import { BluWinged120Spec } from "./blu_winged120";
-import { BluFlame120Spec } from "./blu_flame120";
-import { BluF2PSpec } from "./blu_free_trial";
-import { BluFlame60Spec } from "./blu_flame60";
-import { BluWinged60Spec } from "./blu_winged60";
-import { BluBreath60Spec } from "./blu_breath60";
+import {potRatioSimSpec} from "./common/potency_ratio";
+import { BluWinged120Spec } from "./blu/blu_winged120";
+import { BluFlame120Spec } from "./blu/blu_flame120";
+import { BluF2PSpec } from "./blu/blu_free_trial";
+import { BluFlame60Spec } from "./blu/blu_flame60";
+import { BluWinged60Spec } from "./blu/blu_winged60";
+import { BluBreath60Spec } from "./blu/blu_breath60";
 
 export function registerDefaultSims() {
     registerSim(potRatioSimSpec);
