+++ packages/frontend/src/scripts/sims/healer/sge_sheet_sim.ts
@@ -1,9 +1,9 @@
-import {SimResult, SimSettings, SimSpec, Simulation} from "../simulation";
-import {CharacterGearSet} from "../gear";
+import {SimResult, SimSettings, SimSpec, Simulation} from "../../simulation";
+import {CharacterGearSet} from "../../gear";
 import {applyDhCrit, baseDamage} from "@xivgear/xivmath/xivmath";
 import {ComputedSetStats} from "@xivgear/xivmath/geartypes";
 
-import {FieldBoundCheckBox, labeledCheckbox} from "../components/util";
+import {FieldBoundCheckBox, labeledCheckbox} from "../../components/util";
 
 
 //potencies for our spells
