+++ packages/frontend/src/scripts/sims/healer/sge_sheet_sim_mk2.ts
@@ -1,12 +1,12 @@
-import {SimSettings, SimSpec} from "../simulation";
-import {GcdAbility} from "./sim_types";
+import {SimSettings, SimSpec} from "../../simulation";
+import {GcdAbility} from "../sim_types";
 import {
     BaseMultiCycleSim,
     CycleSimResult,
     ExternalCycleSettings,
     CycleProcessor,
     Rotation
-} from "./sim_processors";
+} from "../sim_processors";
 
 /**
  * Used for all 330p filler abilities
