+++ packages/frontend/src/scripts/simulation.ts
@@ -1,6 +1,6 @@
 import {CharacterGearSet} from "./gear";
 import {JobName, SupportedLevel} from "@xivgear/xivmath/xivconstants";
-import {potRatioSimSpec} from "./sims/potency_ratio";
+import {potRatioSimSpec} from "./sims/common/potency_ratio";
 import {CustomTable} from "./tables";
 import {camel2title} from "./util/strutils";
